requirements.txt
h5py==3.9.0
numpy==1.24.4
Pillow==9.5.0
pytest==7.4.0
python-box==7.0.1
PyYAML==6.0
scikit-learn==1.2.2
scipy==1.10.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
wandb==0.15.4
xgboost==1.7.6